#35 Migrated app to TS
