+++ frontend/src/components/CommanderNavbar.tsx
@@ -1,12 +1,12 @@
 import {Button, Col, Form, Nav, Navbar, Row} from "react-bootstrap";
-import {useAuth} from "../hooks/useAuth.jsx";
+import {useAuth} from "../hooks/useAuth.js";
 
 export const CommanderNavbar = () => {
     const {user, logout} = useAuth();
 
     const userSection = user === null ?
         <></> :
-        <Form inline>
+        <Form>
             <Row>
                 <Col xs="6" />
                 <Col xs="2">
@@ -21,10 +21,10 @@ export const CommanderNavbar = () => {
     return (
         <Navbar expand="lg" className="bg-body-tertiary">
             <Navbar.Brand href="/">
-                <img src="../../public/icon.svg" width="30" height="30" className="d-inline-block align-top"/>
+                <img src="/icon.svg" width="30" height="30" className="d-inline-block align-top" alt="icon"/>
             </Navbar.Brand>
             <Navbar.Toggle aria-controls="basic-navbar-nav"/>
-            <Navbar.Collapse id="basic-navbar-nav">
+            <Navbar.Collapse id="basic-navbar-nav" role="navigation">
                 <Nav className="me-auto">
                     <Nav.Link href="/">Home</Nav.Link>
                     <Nav.Link href="/products">Products</Nav.Link>
