+++ frontend/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import {Container} from "react-bootstrap";
 import {Outlet} from "react-router-dom";
-import {CommanderNavbar} from "./CommanderNavbar.jsx";
+import {CommanderNavbar} from "./CommanderNavbar";
 
 export const Layout = () => (
     <>
