+++ frontend/src/pages/Login.jsx
@@ -1,5 +1,5 @@
 import {useState} from "react";
-import {useAuth} from "../hooks/useAuth.jsx";
+import {useAuth} from "../hooks/useAuth.tsx";
 import {Button, Container, Form} from "react-bootstrap";
 import {useSearchParams} from "react-router-dom";
