+++ frontend/src/types/AuthType.ts
@@ -1,0 +1,5 @@
+export type AuthType = {
+    user: string;
+    login: Function;
+    logout: Function;
+}
